clean up validation
