+++ app/wiring/Controllers.scala
@@ -61,4 +61,10 @@ trait Controllers {
     actionRefiners,
     controllerComponents
   )
+
+  lazy val identityController = new IdentityController(
+    identityService,
+    controllerComponents,
+    actionRefiners
+  )
 }
