+++ assets/components/thankyouPageComponents/thankYou.jsx
