--- assets/helpers/abTests/abtestDefinitions.js.orig
@@ -1,62 +1,0 @@
-// @flow
-import { getQueryParameter } from '../url';
-import type { Tests } from './abtest';
-
-// ----- Tests ----- //
-
-export const tests: Tests = {
-  usRecurringAmountsTestTwo: {
-    variants: ['control', 'higher', 'range'],
-    audiences: {
-      US: {
-        offset: 0,
-        size: 1,
-      },
-    },
-    isActive: true,
-    independent: true,
-    seed: 1,
-  },
-
-  usSecureLogoTest: {
-    variants: ['control', 'logo'],
-    audiences: {
-      US: {
-        offset: 0,
-        size: 1,
-      },
-    },
-    isActive: true,
-    independent: true,
-    seed: 6,
-  },
-
-<<<<<<< HEAD
-  ukDropBottomTest: {
-    variants: ['control', 'no_bottom'],
-    audiences: {
-      GB: {
-        offset: 0,
-        size: 1,
-      },
-    },
-    isActive: true,
-    independent: true,
-    customSegmentCondition: () => getQueryParameter('bundle') === 'contribute',
-=======
-  ukDropBottomTest : {
-    variants: ['control', 'no_bottom'],
-    audiences: {
-        GB: {
-            offset: 0,
-            size: 1,
-        },
-    },
-    isActive: true,
-    independent: true,
->>>>>>> 4ce1e6a846249bcec84be85182615baf42e5ac72
-    seed: 7,
-  },
-
-};
-
