+++ assets/pages/contributions-thankyou/contributionsThankyou.scss
@@ -1,4 +1,4 @@
-#regular-contributions-thankyou-page, #oneoff-contributions-thankyou-page, #regular-contributions-pending-page, #regular-contributions-marketing {
+#regular-contributions-thankyou-page, #oneoff-contributions-thankyou-page, #regular-contributions-pending-page {
 
 	a {
 		color: gu-colour(neutral-1);
