+++ assets/pages/contributions-thankyou/contributionsThankyouActions.js
@@ -14,13 +14,15 @@ function setMarketingPreferencesSelected(): Action {
   return { type: 'MARKETING_PREFERENCES_SELECTED' };
 }
 
+// Fire and forget
 function sendMarketingPreferencesToIdentity(optIn: boolean, email: string): Function {
-  return (dispatch) => {
+  return () => {
     if (optIn) {
       // fire and forget
       return fetch(`${routes.recurringContributionsSendMarketing}?email=${email}`);
     }
-};
+    return null;
+  };
 }
 
 // ----- Exports ----- //
